DELETE :: schedule api
