+++ src/components/main/Schedule/data.ts
@@ -1,0 +1,22 @@
+export const ScheduleItemData = [
+  {
+    id: 1,
+    date: "10월 17일 (화) ~ 20일 (금)",
+    plan: "1차 원서접수",
+  },
+  {
+    id: 2,
+    date: "10월 24일 (화)",
+    plan: "1차 합격자 발표",
+  },
+  {
+    id: 3,
+    date: "10월 28일 (토)",
+    plan: "2차 전형",
+  },
+  {
+    id: 4,
+    date: "11월 2일 (목)",
+    plan: "최종 합격자 발표",
+  },
+];
