+++ src/constants/key.ts
@@ -1,3 +1,2 @@
-export const MAIN_SCHEDULE_LIST = "useMainScheduleList";
 export const MAIN_NOTICE_LIST = "useMainNoticeList";
 export const NOTICE_LIST = "useNoticeList";
