+++ src/features/main.ts
@@ -1,21 +1,6 @@
-import { mainNoticeList, mainScheduleList } from "@/api/main";
+import { mainNoticeList } from "@/api/main";
 import * as KEY from "@/constants/key";
 import { useQuery } from "react-query";
-
-/** 메인 입학 일정 리스트 */
-export interface MainScheduleListType {
-  id: number;
-  date: string;
-  plan: string;
-}
-
-export const useMainScheduleListQuery = () => {
-  const { data } = useQuery<MainScheduleListType[]>(
-    [KEY.MAIN_SCHEDULE_LIST],
-    () => mainScheduleList()
-  );
-  return { data: data || [] };
-};
 
 /** 메인 공지사항 리스트 */
