+++ src/mocks/handlers.ts
@@ -1,5 +1,5 @@
 import { rest } from "msw";
-import { MainNoticeItemData, NoticeItemData, ScheduleItemData } from "./dummy";
+import { MainNoticeItemData, NoticeItemData } from "./dummy";
 
 export const MOCK_API_BASE_URL = "http://localhost:3000";
 
@@ -7,9 +7,7 @@ export const handlers = [
   rest.get(`${MOCK_API_BASE_URL}/main/notice/list`, (req, res, ctx) => {
     return res(ctx.status(200), ctx.json(MainNoticeItemData));
   }),
-  rest.get(`${MOCK_API_BASE_URL}/main/schedule/list`, (req, res, ctx) => {
-    return res(ctx.status(200), ctx.json(ScheduleItemData));
-  }),
+
   rest.get(`${MOCK_API_BASE_URL}/notice/list`, (req, res, ctx) => {
     return res(ctx.status(200), ctx.json(NoticeItemData));
   }),
